fix: domain패키지분리, RacingCarGame에게 의임했던 부분 rollback
