+++ src/main/java/step5/domain/car/CarStrategy.java
@@ -1,4 +1,4 @@
-package step5.car;
+package step5.domain.car;
 
 public interface CarStrategy {
