+++ src/test/java/step5/PreparedCarGameTest.java
@@ -3,13 +3,10 @@ package step5;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step5.car.CarStrategy;
-import step5.game.GameStrategy;
-import step5.game.PreparedCarGame;
-import step5.game.RacingCarGame;
-import step5.message.ExceptionMessage;
-
-import java.util.List;
+import step4.car.CarStrategy;
+import step4.game.PreparedCarGame;
+import step4.game.RacingCarGame;
+import step4.message.ExceptionMessage;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -26,20 +23,18 @@ public class PreparedCarGameTest {
     @Test
     void readyCarTest() {
         int attemptNum = 3;
-        RacingCarGame carGame = spy(new RacingCarGame()); //실체 객체를 mock
-
-        PreparedCarGame preparedCarGame = new PreparedCarGame();
-        preparedCarGame.setGameStrategy(carGame);
-        preparedCarGame.ready("test1, test2, test3", attemptNum);
+        step4.game.RacingCarGame carGame = spy(new step4.game.RacingCarGame()); //실체 객체를 mock
+        carGame.readyRace("test1, test2, test3", attemptNum);
 
         given(carGame.decideForward()) //랜덤으로 전진을 결정하는 메소드를 항상 true로 반환하도록 처리
                 .willReturn(true);
 
+        step4.game.PreparedCarGame preparedCarGame = new step4.game.PreparedCarGame();
+        preparedCarGame.ready(carGame);
+
         preparedCarGame.startRace();
 
-        List<List<CarStrategy>> raceSituation = preparedCarGame.getRaceSituation();
-
-        for (CarStrategy car : raceSituation.get(raceSituation.size() - 1)) {
+        for (CarStrategy car : carGame.getRacingCars()) {
             assertThat(car.getDistance()).isEqualTo(attemptNum);
         }
     }
@@ -49,24 +44,26 @@ public class PreparedCarGameTest {
     @Test
     void checkStartTest() {
         int attemptNum = 3;
+        step4.game.RacingCarGame carGame = new step4.game.RacingCarGame(); //실체 객체를 mock
+        carGame.readyRace("test1, test2, test3", attemptNum);
 
-        PreparedCarGame preparedCarGame = new PreparedCarGame();
-        preparedCarGame.setGameStrategy(new RacingCarGame());
-        preparedCarGame.ready("test1, test2, test3", attemptNum);
+        step4.game.PreparedCarGame preparedCarGame = new step4.game.PreparedCarGame();
+        preparedCarGame.ready(carGame);
 
         assertThatThrownBy(preparedCarGame::winnerRace)
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(ExceptionMessage.GAME_NOT_START.message());
+                .hasMessage(step4.message.ExceptionMessage.GAME_NOT_START.message());
     }
 
     @DisplayName("자동차게임이 이전에 이미 시작되었다면 예외를 발생시킨다.")
     @Test
     void alreadyStartTest() {
         int attemptNum = 3;
+        step4.game.RacingCarGame carGame = new RacingCarGame();
+        carGame.readyRace("test1, test2, test3", attemptNum);
 
-        PreparedCarGame preparedCarGame = new PreparedCarGame();
-        preparedCarGame.setGameStrategy(new RacingCarGame());
-        preparedCarGame.ready("test1, test2, test3", attemptNum);
+        step4.game.PreparedCarGame preparedCarGame = new PreparedCarGame();
+        preparedCarGame.ready(carGame);
 
         preparedCarGame.startRace();
